Refactor file language models and tests
This change replaces the old dict-based file-language response with a typed FileLanguagesMap dataclass and updates imports to use the shared utils package. It also moves category lookup coverage into dedicated tests and updates affected job and API tests to match the new return type, keeping the language metadata API consistent across the app.
