+++ src/main_app/utils/__init__.py
@@ -1,13 +1,13 @@
 """Utility modules for the main application."""
 
-from .verify import verify_required_fields
 from .categories import LANG_CODE_CATEGORY_MAP, lang_code_category
-from .file_langs import get_file_languages, FileLanguages
+from .file_langs import FileLanguagesMap, get_file_languages
+from .verify import verify_required_fields
 
 __all__ = [
     "verify_required_fields",
     "LANG_CODE_CATEGORY_MAP",
     "lang_code_category",
     "get_file_languages",
-    "FileLanguages",
+    "FileLanguagesMap",
 ]
