+++ src/main_app/utils/file_langs.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import logging
-from typing import Any
+from dataclasses import dataclass
 
 import requests
 
@@ -9,133 +9,16 @@ from ..api_services.files_service import get_file_info
 
 logger = logging.getLogger(__name__)
 
-LANG_CODE_CATEGORY_MAP: dict[str, str] = {
-    "ar": "Arabic-language SVG",
-    "bar": "Bavarian-language SVG",
-    "be": "Belarusian-language SVG",
-    "bg": "Bulgarian-language SVG",
-    "bh": "Bhojpuri-language SVG",
-    "bjn": "Banjar-language SVG",
-    "blk": "Pa'O-language SVG",
-    "bn": "Bengali-language SVG",
-    "bs": "Bosnian-language SVG",
-    "ca": "Catalan-language SVG",
-    "ckt": "Chukchi-language SVG",
-    "cs": "Czech-language SVG",
-    "cy": "Welsh-language SVG",
-    "da": "Danish-language SVG",
-    "de": "German-language SVG",
-    "el": "Greek-language SVG",
-    "en": "English-language SVG",
-    "eo": "Esperanto-language SVG",
-    "es": "Spanish-language SVG",
-    "et": "Estonian-language SVG",
-    "eu": "Basque-language SVG",
-    "fa": "Persian-language SVG",
-    "fi": "Finnish-language SVG",
-    "fr": "French-language SVG",
-    "fy": "Frisian-language SVG",
-    "ga": "Irish-language SVG",
-    "gag": "Gagauz-language SVG",
-    "gl": "Galician-language SVG",
-    "got": "Gothic-language SVG",
-    "gu": "Gujarati-language SVG",
-    "he": "Hebrew-language SVG",
-    "hi": "Hindi-language SVG",
-    "hr": "Croatian-language SVG",
-    "hsb": "Sorbian-language SVG",
-    "ht": "Haitian Creole-language SVG",
-    "hu": "Hungarian-language SVG",
-    "ia": "Interlingua-language SVG",
-    "id": "Indonesian-language SVG",
-    "is": "Icelandic-language SVG",
-    "isv": "Interslavic-language SVG",
-    "it": "Italian-language SVG",
-    "ja": "Japanese-language SVG",
-    "jv": "Javanese-language SVG",
-    "ka": "Georgian-language SVG",
-    "kk": "Kazakh-language SVG",
-    "kn": "Kannada-language SVG",
-    "ko": "Korean-language SVG",
-    "ks": "Kashmiri-language SVG",
-    "ksh": "Ripuarian-language SVG",
-    "ku": "Kurdish-language SVG",
-    "ky": "Kyrgyz-language SVG",
-    "la": "Latin-language SVG",
-    "lb": "Luxembourgish-language SVG",
-    "lrc": "Luri-language SVG",
-    "lt": "Lithuanian-language SVG",
-    "lv": "Latvian-language SVG",
-    "mag": "Magahi-language SVG",
-    "mai": "Maithili-language SVG",
-    "mdf": "Moksha-language SVG",
-    "mg": "Malagasy-language SVG",
-    "mi": "Maori-language SVG",
-    "mk": "Macedonian-language SVG",
-    "ml": "Malayalam-language SVG",
-    "mn": "Mongolian-language SVG",
-    "mr": "Marathi-language SVG",
-    "mrj": "Hill Mari-language SVG",
-    "ms": "Malay-language SVG",
-    "mt": "Maltese-language SVG",
-    "my": "Burmese-language SVG",
-    "nds": "Low German-language SVG",
-    "ne": "Nepali-language SVG",
-    "nl": "Dutch-language SVG",
-    "no": "Norwegian-language SVG",
-    "nrm": "Norman-language SVG",
-    "pa": "Punjabi-language SVG",
-    "pl": "Polish-language SVG",
-    "pms": "Piedmontese-language SVG",
-    "prs": "Dari-language SVG",
-    "pt": "Portuguese-language SVG",
-    "qu": "Quechua-language SVG",
-    "rm": "Romansh-language SVG",
-    "ro": "Romanian-language SVG",
-    "ru": "Russian-language SVG",
-    "rue": "Rusyn-language SVG",
-    "sa": "Sanskrit-language SVG",
-    "sat": "Santali-language SVG",
-    "sco": "Scots-language SVG",
-    "sd": "Sindhi-language SVG",
-    "se": "Northern Sami-language SVG",
-    "sh": "Serbo-Croatian-language SVG",
-    "si": "Sinhala-language SVG",
-    "sk": "Slovak-language SVG",
-    "sl": "Slovene-language SVG",
-    "smn": "Inari Sami-language SVG",
-    "sms": "Skolt Sami-language SVG",
-    "so": "Somali-language SVG",
-    "sr": "Serbian-language SVG",
-    "sv": "Swedish-language SVG",
-    "sw": "Swahili-language SVG",
-    "syl": "Sylheti-language SVG",
-    "ta": "Tamil-language SVG",
-    "te": "Telugu-language SVG",
-    "th": "Thai-language SVG",
-    "tl": "Tagalog-language SVG",
-    "tr": "Turkish-language SVG",
-    "tt": "Tatar-language SVG",
-    "tup": "Tupi-language SVG",
-    "udm": "Udmurt-language SVG",
-    "uk": "Ukrainian-language SVG",
-    "ur": "Urdu-language SVG",
-    "uz": "Uzbek-language SVG",
-    "vec": "Venetian-language SVG",
-    "vi": "Vietnamese-language SVG",
-    "wa": "Walloon-language SVG",
-    "yi": "Yiddish-language SVG",
-    "yo": "Yoruba-language SVG",
-    "zh": "Chinese-language SVG",
-    "zu": "Zulu-language SVG",
-}
+
+@dataclass
+class FileLanguagesMap:
+    """Result of extracting available SVG translation languages for a Commons file."""
+
+    error: str | None
+    langs: list[str] | None
 
 
-def lang_code_category(langcode: str) -> str | None:
-    return LANG_CODE_CATEGORY_MAP.get(langcode)
-
-
-def get_file_languages(file_name: str, session: requests.Session | None = None) -> dict[str, Any]:
+def get_file_languages(file_name: str, session: requests.Session | None = None) -> FileLanguagesMap:
     """
     Extract available SVG translation languages for a given Commons file.
 
@@ -144,12 +27,12 @@ def get_file_languages(file_name: str, session: requests.Session | None = None) 
         session: Optional pre-configured requests session.
 
     Returns:
-        Dictionary containing an 'error' message (if any) and a 'langs' list.
+        A `FileLanguagesMap` instance with `error` (if any) and `langs` list.
     To mirror:
         https://svgtranslate.toolforge.org/api/languages/File:Parkinsons_disease_prevalence_ihme,_Africa,_2021.svg
     """
     if not file_name:
-        return {"error": "Empty fileName", "langs": None}
+        return FileLanguagesMap(error="Empty fileName", langs=None)
 
     # Normalize file name by stripping leading "File:" prefix
     file_name = file_name.strip()
@@ -159,17 +42,20 @@ def get_file_languages(file_name: str, session: requests.Session | None = None) 
     file_info = get_file_info(prefixed_file_name, session=session)
     # check if image exists
     if file_info.exists is False:
-        return {"error": "File does not exist", "langs": None}
+        return FileLanguagesMap(error="File does not exist", langs=None)
 
     # Extract metadata array
     imageinfo = file_info.imageinfo
     if not imageinfo:
-        return {"error": f"Metadata not found for {prefixed_file_name}. Error: {file_info.error}", "langs": None}
+        return FileLanguagesMap(
+            error=f"Metadata not found for {prefixed_file_name}. Error: {file_info.error}",
+            langs=None,
+        )
 
     # metadata shema: [ { "name": "version", "value": 2 }, ... , { "name": "translations", "value": []} ]
     metadata = imageinfo[0].get("metadata", [])
     if not metadata:
-        return {"error": f"Metadata array empty for {prefixed_file_name}", "langs": None}
+        return FileLanguagesMap(error=f"Metadata array empty for {prefixed_file_name}", langs=None)
 
     translations = []
     for x in metadata:
@@ -181,12 +67,12 @@ def get_file_languages(file_name: str, session: requests.Session | None = None) 
     if isinstance(translations, list) and len(translations) > 0:
         # Extract language codes from translation entries
         langs_keys = [t["name"] for t in translations if isinstance(t, dict) and "name" in t]
-        return {"error": None, "langs": langs_keys if langs_keys else ["en"]}
+        return FileLanguagesMap(error=None, langs=langs_keys if langs_keys else ["en"])
 
-    return {"error": None, "langs": ["en"]}
+    return FileLanguagesMap(error=None, langs=["en"])
+
 
 __all__ = [
-    "LANG_CODE_CATEGORY_MAP",
+    "FileLanguagesMap",
     "get_file_languages",
-    "lang_code_category",
 ]
